+++ src/pages/generate/components/result/ResultPage.tsx
@@ -12,7 +12,7 @@ import {
   useResultData,
 } from '../../hooks/useGenerate';
 import type { GenerateImageData } from '../../types/GenerateType';
-import type { MyPageImageDetailData } from '@/pages/mypage/types/apis/MyPageType';
+import type { MyPageImageDetailData } from '@/pages/mypage/types/apis/MyPage';
 import { useMyPageImageDetail } from '@/pages/mypage/hooks/useMypage';
 import LikeButton from '@/shared/components/button/likeButton/LikeButton';
 import DislikeButton from '@/shared/components/button/likeButton/DislikeButton';
